render input as list

+++ src/session.rs
@@ -2,5 +2,5 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Session {
-    pub depth: u16,
+    pub depth: usize,
 }
